change directory items.js
